twitch-bot/requirements.txt
twitchio==2.9.1
websockets
python-dotenv
asyncio